+++ R/reshape_msd.R
@@ -1,7 +1,8 @@
 #' Reshape Semi-Wide MSD
 #'
 #' @param df MSD dataset in the semi-wide format
-#' @param direction direction of reshape, wide (default) or long
+#' @param direction direction of reshape, long (default) or wide (original MSD structure)
+#' @param clean clean period for graphing, eg(fy2019qtr2 -> FY19Q2) and create a period type (targets, results, cumulative)
 #'
 #' @export
 #'
@@ -11,29 +12,44 @@
 #'   df_genie <- match_msd("~/Downloads/PEPFAR-Data-Genie-PSNUByIMs-2018-08-15.zip")
 #'  #reshape long
 #'   df_genie_long <- reshape_msd(df_genie, direction = "long")
+#'  #reshape long + clean
+#'   df_genie_long <- reshape_msd(df_genie, direction = "long", clean = TRUE)
 #'  #or reshape wide
 #'   df_genie_wide <- reshape_msd(df_genie, direction = "wide") }
 
-reshape_msd <- function(df, direction = c("wide", "long")){
+reshape_msd <- function(df, direction = c("long", "wide"), clean = FALSE){
 
   #check if upper case (for FY or fy names)
     is_upper <- stringr::str_detect(names(df)[1], "[[:upper:]]")
+    fy_var <- ifelse(is_upper, dplyr::sym("Fiscal_Year"), dplyr::sym("fiscal_year"))
 
   #reshape long (wide need to be reshaped long first as well)
     df <- df %>%
-      tidyr::gather(period, val, targets:cumulative, na.rm = TRUE) %>%
+      tidyr::gather(period, val,
+                    dplyr::matches("TARGETS|targets|(Q|q)tr|(C|c)umulative"),
+                    na.rm = TRUE) %>%
       dplyr::filter(val != 0) %>%
       dplyr::mutate(period = stringr::str_remove(period, "tr"), #remove "tr" from "Qtr" to match old
                     period = stringr::str_replace(period, "(TARGETS|targets)", "_\\1"), #add _ to match old
-                    fiscal_year = paste0(ifelse(is_upper, "FY", "fy"), fiscal_year)) %>%  #add FY to match old
-      tidyr::unite(period, c(fiscal_year, period), sep = "") #combine fy and pd together
+                    !!fy_var := paste0(ifelse(is_upper, "FY", "fy"), !!fy_var)) %>%  #add FY to match old
+      tidyr::unite(period, c(!!fy_var, period), sep = "") #combine fy and pd together
 
-  #reshape wide (default)
-    if("wide" %in% direction){
+  #reshape wide
+    if(!"long" %in% direction){
       df <- df %>%
         dplyr::mutate(period = stringr::str_replace(period, "(C|c)um", "zzz.\\1um")) %>% #add z to reorder correctly
         tidyr::spread(period, val) %>%
         dplyr::rename_all( ~ stringr::str_remove(.,"zzz.")) #remove zzz
+    }
+
+  #clean
+    if("long" %in% direction && clean == TRUE){
+      df <- df %>%
+        dplyr::mutate(period_type = stringr::str_extract(period, "TARGETS|targets|(C|c)umulative") %>% tolower,
+                      period_type = ifelse(is.na(period_type), "results", period_type),
+                      period = stringr::str_remove(period, "20") %>% toupper,
+                      period = stringr::str_remove(period, "CUMULATIVE|_TARGETS")) %>%
+        dplyr::select(-val, dplyr::everything())
     }
 
   return(df)